move hourly bar legends functions into plot_settings.py
